Code review fixes and more
