+++ src/containers/NetworkPanel/NetworkDetails/TracingView.tsx
@@ -1,7 +1,7 @@
-import { TracingVisualization } from "../../../components/TracingVisualization";
+import { TracingVisualization } from "@/components/TracingVisualization";
 import { IResponseBody } from "@/types";
 import React, { useMemo } from "react";
-import * as safeJson from "../../../helpers/safeJson";
+import * as safeJson from "@/helpers/safeJson";
 
 interface ITracingViewProps {
   response?: string;
